requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
msoffcrypto-tool==5.0.0
plotly==5.18.0
xlrd==2.0.1
